Refactor HtmlParser
Refactored the HtmlParser to address issues with the parsing flow and to cover additional elements - spans, headers, and raw nodes have been introduced to the parsing process. The HtmlParser now takes a single argument of a `NodeFactory` instance, instead of passing through the NodeFactory instance in every method. This significantly optimizes the parse flow and makes the code more readable. These changes have also been factored into the `DefaultMarkdownParserModule`.